requirements.txt
aiohttp==1.0.5
async-timeout==1.2.0
cffi==1.10.0
chardet==2.3.0
discord.py==0.16.7
gTTS==1.1.8
gTTS-token==1.1.1
multidict==2.1.4
pycparser==2.17
PyNaCl==1.1.2
regex==2017.4.5
requests==2.13.0
six==1.10.0
websockets==3.3
youtube-dl==2017.4.3
